python_api/requirements.txt
fastapi
uvicorn
python-multipart
mediapipe
numpy
Pillow
